feat(graphql): rm old api, add graphql endpoint
